[CommunicationIdentifier] implement new global method to createCommunicationIdentifier from rawId
