+++ AzureSDK.xcworkspace/xcshareddata/xcschemes/communication.xcscheme
@@ -86,7 +86,7 @@
       codeCoverageEnabled = "YES">
       <Testables>
          <TestableReference
-            skipped = "NO">
+            skipped = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "D1B7EAEC257F02FC004F384A"
@@ -96,7 +96,7 @@
             </BuildableReference>
          </TestableReference>
          <TestableReference
-            skipped = "NO">
+            skipped = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "D147E62725CE242C001CFB5D"
