+++ sdk/communication/AzureCommunicationCommon/Tests/CommunicationIdentifierTest.swift
@@ -150,8 +150,141 @@ class CommunicationIdentifierTest: XCTestCase {
                 )
         )
     }
-
     // swiftlint:enable function_body_length
+
+    func test_createUnknownIdentifier() {
+        var unknownIdentifier = createCommunicationIdentifier(from: "37691ec4-57fb-4c0f-ae31-32791610cb14")
+        XCTAssertTrue(unknownIdentifier.isKind(of: UnknownIdentifier.self))
+        XCTAssertEqual(unknownIdentifier.kind, .unknown)
+        XCTAssertEqual(unknownIdentifier.rawId, "37691ec4-57fb-4c0f-ae31-32791610cb14")
+
+        unknownIdentifier = createCommunicationIdentifier(from: "48:37691ec4-57fb-4c0f-ae31-32791610cb14")
+        XCTAssertTrue(unknownIdentifier.isKind(of: UnknownIdentifier.self))
+        XCTAssertEqual(unknownIdentifier.rawId, "48:37691ec4-57fb-4c0f-ae31-32791610cb14")
+    }
+
+    func test_createPhoneNumberIdentifier() {
+        let phoneNumberRawId = "4:12345556789"
+        let phoneNumberIdentifier = createCommunicationIdentifier(from: phoneNumberRawId)
+        XCTAssertTrue(phoneNumberIdentifier.isKind(of: PhoneNumberIdentifier.self))
+        XCTAssertEqual(phoneNumberIdentifier.kind, .phoneNumber)
+        XCTAssertEqual((phoneNumberIdentifier as? PhoneNumberIdentifier)?.rawId, phoneNumberRawId)
+        XCTAssertEqual((phoneNumberIdentifier as? PhoneNumberIdentifier)?.phoneNumber, "+12345556789")
+    }
+
+    func test_createCommunicationUserIdentifier() {
+        let acsRawId = "8:acs:37691ec4-57fb-4c0f-ae31-32791610cb14_37691ec4-57fb-4c0f-ae31-32791610cb14"
+        var communicationUserIdentifier = createCommunicationIdentifier(from: acsRawId)
+        XCTAssertTrue(communicationUserIdentifier.isKind(of: CommunicationUserIdentifier.self))
+        XCTAssertEqual(communicationUserIdentifier.kind, .communicationUser)
+        XCTAssertEqual(communicationUserIdentifier.rawId, acsRawId)
+        XCTAssertEqual(
+            (communicationUserIdentifier as? CommunicationUserIdentifier)?.identifier, acsRawId
+        )
+
+        let spoolRawId = "8:spool:37691ec4-57fb-4c0f-ae31-32791610cb14_37691ec4-57fb-4c0f-ae31-32791610cb14"
+        communicationUserIdentifier = createCommunicationIdentifier(from: spoolRawId)
+        XCTAssertTrue(communicationUserIdentifier.isKind(of: CommunicationUserIdentifier.self))
+        XCTAssertEqual(communicationUserIdentifier.kind, .communicationUser)
+        XCTAssertEqual(communicationUserIdentifier.rawId, spoolRawId)
+        XCTAssertEqual(
+            (communicationUserIdentifier as? CommunicationUserIdentifier)?.identifier, spoolRawId
+        )
+
+        let dodAcsRawId = "8:dod-acs:37691ec4-57fb-4c0f-ae31-32791610cb14_37691ec4-57fb-4c0f-ae31-32791610cb14"
+        communicationUserIdentifier = createCommunicationIdentifier(from: dodAcsRawId)
+        XCTAssertTrue(communicationUserIdentifier.isKind(of: CommunicationUserIdentifier.self))
+        XCTAssertEqual(communicationUserIdentifier.kind, .communicationUser)
+        XCTAssertEqual(communicationUserIdentifier.rawId, dodAcsRawId)
+        XCTAssertEqual(
+            (communicationUserIdentifier as? CommunicationUserIdentifier)?.identifier, dodAcsRawId
+        )
+
+        let gcchAcsRawId = "8:gcch-acs:37691ec4-57fb-4c0f-ae31-32791610cb14_37691ec4-57fb-4c0f-ae31-32791610cb14"
+        communicationUserIdentifier = createCommunicationIdentifier(from: gcchAcsRawId)
+        XCTAssertTrue(communicationUserIdentifier.isKind(of: CommunicationUserIdentifier.self))
+        XCTAssertEqual(communicationUserIdentifier.kind, .communicationUser)
+        XCTAssertEqual(communicationUserIdentifier.rawId, gcchAcsRawId)
+        XCTAssertEqual(
+            (communicationUserIdentifier as? CommunicationUserIdentifier)?.identifier, gcchAcsRawId
+        )
+    }
+
+    func test_createMicrosoftTeamsUserIdentifierAnonymousScope() {
+        let teamUserAnonymousScope =
+            "8:teamsvisitor:37691ec4-57fb-4c0f-ae31-32791610cb14_37691ec4-57fb-4c0f-ae31-32791610cb14"
+        let teamUserIdentifier = createCommunicationIdentifier(from: teamUserAnonymousScope)
+        XCTAssertTrue(teamUserIdentifier.isKind(of: MicrosoftTeamsUserIdentifier.self))
+        XCTAssertEqual(teamUserIdentifier.kind, .microsoftTeamsUser)
+        XCTAssertEqual(teamUserIdentifier.rawId, teamUserAnonymousScope)
+        XCTAssertEqual(
+            (teamUserIdentifier as? MicrosoftTeamsUserIdentifier)?.userId,
+            "37691ec4-57fb-4c0f-ae31-32791610cb14_37691ec4-57fb-4c0f-ae31-32791610cb14"
+        )
+        XCTAssertEqual(
+            (teamUserIdentifier as? MicrosoftTeamsUserIdentifier)?.isAnonymous, true
+        )
+        XCTAssertEqual(
+            (teamUserIdentifier as? MicrosoftTeamsUserIdentifier)?.cloudEnviroment, .Public
+        )
+    }
+
+    func test_createMicrosoftTeamsUserIdentifierPublicScope() {
+        let teamUserPublicCloudScope =
+            "8:orgid:37691ec4-57fb-4c0f-ae31-32791610cb14_37691ec4-57fb-4c0f-ae31-32791610cb14"
+        let teamUserIdentifier = createCommunicationIdentifier(from: teamUserPublicCloudScope)
+        XCTAssertTrue(teamUserIdentifier.isKind(of: MicrosoftTeamsUserIdentifier.self))
+        XCTAssertEqual(teamUserIdentifier.kind, .microsoftTeamsUser)
+        XCTAssertEqual(teamUserIdentifier.rawId, teamUserPublicCloudScope)
+        XCTAssertEqual(
+            (teamUserIdentifier as? MicrosoftTeamsUserIdentifier)?.userId,
+            "37691ec4-57fb-4c0f-ae31-32791610cb14_37691ec4-57fb-4c0f-ae31-32791610cb14"
+        )
+        XCTAssertEqual(
+            (teamUserIdentifier as? MicrosoftTeamsUserIdentifier)?.isAnonymous, false
+        )
+        XCTAssertEqual(
+            (teamUserIdentifier as? MicrosoftTeamsUserIdentifier)?.cloudEnviroment, .Public
+        )
+    }
+
+    func test_createMicrosoftTeamsUserIdentifierDODScope() {
+        let teamUserDODCloudScope = "8:dod:37691ec4-57fb-4c0f-ae31-32791610cb14_37691ec4-57fb-4c0f-ae31-32791610cb14"
+        let teamUserIdentifier = createCommunicationIdentifier(from: teamUserDODCloudScope)
+        XCTAssertTrue(teamUserIdentifier.isKind(of: MicrosoftTeamsUserIdentifier.self))
+        XCTAssertEqual(teamUserIdentifier.kind, .microsoftTeamsUser)
+        XCTAssertEqual(teamUserIdentifier.rawId, teamUserDODCloudScope)
+        XCTAssertEqual(
+            (teamUserIdentifier as? MicrosoftTeamsUserIdentifier)?.userId,
+            "37691ec4-57fb-4c0f-ae31-32791610cb14_37691ec4-57fb-4c0f-ae31-32791610cb14"
+        )
+        XCTAssertEqual(
+            (teamUserIdentifier as? MicrosoftTeamsUserIdentifier)?.isAnonymous, false
+        )
+        XCTAssertEqual(
+            (teamUserIdentifier as? MicrosoftTeamsUserIdentifier)?.cloudEnviroment, .Dod
+        )
+    }
+
+    func test_createMicrosoftTeamsUserIdentifierGCCHScope() {
+        let teamUserGCCHCloudScope = "8:gcch:37691ec4-57fb-4c0f-ae31-32791610cb14_37691ec4-57fb-4c0f-ae31-32791610cb14"
+        let teamUserIdentifier = createCommunicationIdentifier(from: teamUserGCCHCloudScope)
+
+        XCTAssertTrue(teamUserIdentifier.isKind(of: MicrosoftTeamsUserIdentifier.self))
+        XCTAssertEqual(teamUserIdentifier.kind, .microsoftTeamsUser)
+        XCTAssertEqual(teamUserIdentifier.rawId, teamUserGCCHCloudScope)
+
+        XCTAssertEqual(
+            (teamUserIdentifier as? MicrosoftTeamsUserIdentifier)?.userId,
+            "37691ec4-57fb-4c0f-ae31-32791610cb14_37691ec4-57fb-4c0f-ae31-32791610cb14"
+        )
+        XCTAssertEqual(
+            (teamUserIdentifier as? MicrosoftTeamsUserIdentifier)?.isAnonymous, false
+        )
+        XCTAssertEqual(
+            (teamUserIdentifier as? MicrosoftTeamsUserIdentifier)?.cloudEnviroment, .Gcch
+        )
+    }
 
     func test_MicrosoftTeamsUserIdentifier_DefaultCloudIsPublic() throws {
         XCTAssertEqual(
